Try using BCLK as clock source for DAC.
The clocks of the DSP and the DAC should have the same source, but they
do not. So using BCLK of the I2S as master clock source for the DAC
could solve this issue. It did not work yet, as the PLL did not lock.
These needs to be investigated further, maybe the DAC has to be powered
down for a while?