DAC works now :party:
So the guy from Analog Devices was right: Different master clocks is the
issue. In the last commit my temporary fix didn't work because I
connected the BCLK from DSP to XTALO, but not MCLKI on the DAC. The
audio sounds good right now, but a more long-term solution has to be
found. Either it is possible to resolve the drifting clock issue in
software (make the DSP reference the DAC) or a connection with a
soldered wire has to be made.