Some adjustments for audio output
The PCM1865 signal now goes through an ASRC too, as in theory the clocks
may also drift. A crossover for the subwoofer channel was also added
with 3db gain, as the subwoofer has a lower SPL/1W than my other
speakers. The TPA3116D2 amplifier ICs are still not noise-free, the
power supply is probably not completely free noise. It couples into the
audio path somehow. The noise is louder when I use my SMPS, and not as
loud but still present with my lab bench power supply. The noise is only
audible at very low volume.